pyproject.toml
[project]
name = "langsmith-course"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.9"
dependencies = [
    "anthropic>=0.69.0",
    "langchain-community>=0.3.30",
    "langchain-core>=0.3.78",
    "langchain-huggingface>=0.3.1",
    "langchain-openai>=0.3.34",
    "langchain-voyageai>=0.1.7",
    "langgraph>=0.6.8",
    "langgraph-checkpoint-sqlite>=2.0.11",
    "langgraph-sdk>=0.2.9",
    "langsmith>=0.2.0",
    "lxml>=6.0.2",
    "notebook>=7.4.7",
    "pandas>=2.3.3",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.6.1",
    "sentence-transformers>=5.1.1",
]
